Support for over-riding the DB in flask config.
Flask MongoKit is restrictive in the sense that it assumes that the
database configured in the app.config would be the only database used.
This patch proposes to allow usage of multiple MongoDB Databases in the
same flask by providing the database parameter as a part of the
constructor.

```python
    foo_db = MongoKit(database='foo_db')
```